Automate tests using \showoutput
This needs burning-edge l3build, and at least at present also a
locally-modifed biblatex.sty with valid values for version/date.

The idea is that this allows the example files to be used exactly as
they are. As noted in the previous commit, the test outcomes still need
checking.

A few more l3build changes should allow it to build biblatex entirely:
the main issue is the versioning, with perhaps "flatten" needing
extending to the TDS form.